fix: files clean up
