+++ Notes.md
@@ -1,3 +1,5 @@
-Hey, this files will be for notes and anything we want to keep track of. 
+Hey, this files will be for notes and anything we want to keep track of.
 
 - For temp basis I have created a folder with name `data` to test things out.
+
+- Created shared folder for common assets being used by different functions
