--- data/Pradumna.json
@@ -1,7 +1,0 @@
-{
-  "name": "a",
-  "type": "Personal",
-  "bio": "asas",
-  "links": [],
-  "milestones": []
-}
