+++ data/Pradumnasaraf.json
@@ -1,0 +1,15 @@
+{
+  "name": "Prad",
+  "type": "Personal",
+  "bio": "dev",
+  "links": [
+    {
+      "name": "git",
+      "url": "jfhjhgjf",
+      "icon": "FaGithub"
+    }
+  ],
+  "testimonials": [
+    "eddiejaoude"
+  ]
+}
