+++ src/addevent/addEvent.js
@@ -1,0 +1,3 @@
+const addEvent = () => {};
+
+module.exports = addEvent;
