+++ src/shared/checkUser.js
