+++ src/shared/questions/links.js
@@ -1,0 +1,88 @@
+const { prompt } = require("enquirer");
+const { geticons } = require("../assets/icons");
+let links = [];
+let icons = [];
+
+async function addlinks(bool) {
+  icons = await geticons();
+
+  while (bool) {
+    let answers = await prompt([
+      {
+        type: "input",
+        name: "name",
+        message: "What is the name of the link?",
+      },
+      {
+        type: "input",
+        name: "url",
+        message: "What is the URL of the link?",
+      },
+      {
+        type: "select",
+        name: "icon",
+        choices: icons,
+        message: "Choose an icon (Press down arrow to see more options)",
+      },
+      {
+        type: "confirm",
+        name: "addLink",
+        message: "Do you want to add another link?",
+      },
+    ]);
+    links.push({
+      name: answers.name,
+      url: answers.url,
+      icon: answers.icon,
+    });
+    if (!answers.addLink) {
+      return links;
+    }
+  }
+}
+
+async function removelinks(links) {
+  let choiceLinks = links;
+  let stop = false;
+  while (!stop) {
+    const answers = await prompt([
+      {
+        type: "select",
+        name: "link",
+        choices: choiceLinks,
+        message: "Choose which one you want to remove",
+      },
+    ]);
+    choiceLinks = choiceLinks.filter((link) => link.name !== answers.link);
+    if (choiceLinks.length !== 0) {
+      const res = await prompt([
+        {
+          type: "confirm",
+          name: "removeLink",
+          message: "Do you want to remove another link?",
+        },
+      ]);
+      if (!res.removeLink) {
+        let result = [];
+        choiceLinks.map((link) => {
+          const { name, url, icon } = link;
+          result.push({ name, url, icon });
+        });
+        return result;
+      }
+    } else {
+      return [];
+    }
+  }
+}
+
+function updatelinks(links) {
+  console.log("updating links");
+  return links;
+}
+
+module.exports = {
+  addlinks,
+  removelinks,
+  updatelinks,
+};
