+++ src/shared/questions/testimonials.js
@@ -1,0 +1,71 @@
+const { prompt } = require("enquirer");
+let testimonials = [];
+
+async function addtestimonials(bool) {
+  while (bool) {
+    let answers = await prompt([
+      {
+        type: "input",
+        name: "name",
+        message: "Who gave the testimonial?(write their username)",
+      },
+      {
+        type: "confirm",
+        name: "addTestimonial",
+        message: "Do you want to add another testimonial?",
+      },
+    ]);
+    testimonials.push(answers.name);
+    if (!answers.addTestimonial) {
+      return testimonials;
+    }
+  }
+}
+
+async function removetestimonials(testimonials) {
+  let choiceTestimonials = testimonials;
+  let stop = false;
+  while (!stop) {
+    const answers = await prompt([
+      {
+        type: "select",
+        name: "testimonial",
+        choices: choiceTestimonials,
+        message: "Choose which one you want to remove",
+      },
+    ]);
+    choiceTestimonials = choiceTestimonials.filter(
+      (testimonial) => testimonial.name !== answers.testimonial
+    );
+    if (choiceTestimonials.length !== 0) {
+      const res = await prompt([
+        {
+          type: "confirm",
+          name: "removeTestimonial",
+          message: "Do you want to remove another testimonial?",
+        },
+      ]);
+      if (!res.removeTestimonial) {
+        let result = [];
+        choiceTestimonials.map((testimonial) => {
+          const { name } = testimonial;
+          result.push(name);
+        });
+        return result;
+      }
+    } else {
+      return [];
+    }
+  }
+}
+
+async function updatetestimonials(testimonials) {
+  console.log("updating testimonials");
+  return testimonials;
+}
+
+module.exports = {
+  addtestimonials,
+  removetestimonials,
+  updatetestimonials,
+};
